refactor: rename Interfaces syntax to Contracts
