+++ config/auto-doc.php
@@ -110,7 +110,7 @@ return [
     |
     | The name of driver, which will collect and save documentation
     | Feel free to use your own driver class which should be inherited from
-    | `RonasIT\AutoDoc\Interfaces\SwaggerDriverInterface` interface,
+    | `RonasIT\AutoDoc\Contracts\SwaggerDriverContract` interface,
     | or one of our drivers from the `drivers` config:
     */
     'driver' => env('SWAGGER_DRIVER', 'local'),
