+++ src/Contracts/SwaggerDriverContract.php
@@ -1,8 +1,8 @@
 <?php
 
-namespace RonasIT\AutoDoc\Interfaces;
+namespace RonasIT\AutoDoc\Contracts;
 
-interface SwaggerDriverInterface
+interface SwaggerDriverContract
 {
     /**
      * Save temporary data
